refactor(sidecar): split pass-utils by concern
Export moves:
- VueScriptBlock -> src/vue-script.ts
- extractVueScripts -> VueScript.extractBlocks in src/vue-script.ts
- scriptAttribute -> VueScript.attribute in src/vue-script.ts
- isJavaScriptOrTypeScript -> VueScript.isJavaScriptOrTypeScript in src/vue-script.ts
- CallParens -> src/source-text.ts
- lineStart -> SourceText.lineStart in src/source-text.ts
- lineIndent -> SourceText.lineIndent in src/source-text.ts
- sourceOf -> SourceText.sourceOf in src/source-text.ts
- hasCommentBetween -> SourceText.hasCommentBetween in src/source-text.ts
- callParens -> SourceText.callParens in src/source-text.ts
- unwrapChainExpression -> SourceText.unwrapChainExpression in src/source-text.ts
- rangesOverlap -> Edits.rangesOverlap in src/edits.ts
- nonOverlappingEdits -> Edits.nonOverlapping in src/edits.ts
- isTargetFile -> src/file-targets.ts
- isDeclarationFile -> src/file-targets.ts
- isConstDeclaration -> src/ast.ts

Sources.parse and NodeSourceFiles already own parsing and atomic file I/O, so no related pass-utils exports remained.

Assessment: drizzle-queries has clear detection and emission phases, now marked with section comments. Keeping the tightly coupled helpers in one file avoids a subdirectory split.
