+++ packages/ts/sidecar/package.json
@@ -26,8 +26,8 @@
 		"#sidecar/format-all.js": "./src/format-all.ts",
 		"#sidecar/format-pipeline": "./src/format-pipeline.ts",
 		"#sidecar/format-pipeline.js": "./src/format-pipeline.ts",
-		"#sidecar/pass-utils": "./src/pass-utils.ts",
-		"#sidecar/pass-utils.js": "./src/pass-utils.ts",
+		"#sidecar/file-targets": "./src/file-targets.ts",
+		"#sidecar/file-targets.js": "./src/file-targets.ts",
 		"#sidecar/files": "./src/files.ts",
 		"#sidecar/files.js": "./src/files.ts",
 		"#sidecar/fluent-chains": "./src/fluent-chains.ts",
@@ -42,9 +42,13 @@
 		"#sidecar/segment.js": "./src/segment.ts",
 		"#sidecar/source-files": "./src/source-files.ts",
 		"#sidecar/source-files.js": "./src/source-files.ts",
+		"#sidecar/source-text": "./src/source-text.ts",
+		"#sidecar/source-text.js": "./src/source-text.ts",
 		"#sidecar/sources": "./src/sources.ts",
 		"#sidecar/sources.js": "./src/sources.ts",
 		"#sidecar/types": "./src/types.ts",
+		"#sidecar/vue-script": "./src/vue-script.ts",
+		"#sidecar/vue-script.js": "./src/vue-script.ts",
 		"#sidecar/validate-syntax": "./src/validate-syntax.ts",
 		"#sidecar/validate-syntax.js": "./src/validate-syntax.ts"
 	},
