+++ packages/ts/sidecar/src/package.json
@@ -25,8 +25,8 @@
 		"#sidecar/format-all.js": "./format-all.ts",
 		"#sidecar/format-pipeline": "./format-pipeline.ts",
 		"#sidecar/format-pipeline.js": "./format-pipeline.ts",
-		"#sidecar/pass-utils": "./pass-utils.ts",
-		"#sidecar/pass-utils.js": "./pass-utils.ts",
+		"#sidecar/file-targets": "./file-targets.ts",
+		"#sidecar/file-targets.js": "./file-targets.ts",
 		"#sidecar/files": "./files.ts",
 		"#sidecar/files.js": "./files.ts",
 		"#sidecar/fluent-chains": "./fluent-chains.ts",
@@ -41,9 +41,13 @@
 		"#sidecar/segment.js": "./segment.ts",
 		"#sidecar/source-files": "./source-files.ts",
 		"#sidecar/source-files.js": "./source-files.ts",
+		"#sidecar/source-text": "./source-text.ts",
+		"#sidecar/source-text.js": "./source-text.ts",
 		"#sidecar/sources": "./sources.ts",
 		"#sidecar/sources.js": "./sources.ts",
 		"#sidecar/types": "./types.ts",
+		"#sidecar/vue-script": "./vue-script.ts",
+		"#sidecar/vue-script.js": "./vue-script.ts",
 		"#sidecar/validate-syntax": "./validate-syntax.ts",
 		"#sidecar/validate-syntax.js": "./validate-syntax.ts"
 	}
